Passwords hashing
+ added password hashing during registration
+ extracted StudentUserDetailsService because of cycle dependencies
